finish the parser, and test it properly
Bring the parser to the same standard as the lexer and cover it: the
suite goes from 26 cases to 57.

The new tests exercise precedence and associativity, prefix operators,
grouping, function calls, member access, array literals, equality
operators, and printing an AST back to a form that lexes again. Several
fixes to precedences and to ast_print fall out of what they exposed, and
exponentiation is now right-associative, so 2**3**2 is 2**(3**2) as it is
everywhere else.

Numeric literals are held at double precision. As float, a threshold did
not survive being parsed:

  273.15 as float  = 273.14999389648438
  273.15 as double = 273.14999999999998

and a legal value like 1e40 was rejected outright as out of range. For a
language whose main job is expressing thresholds over climate fields,
narrowing in the AST is the wrong place to do it; a consumer that wants
single precision can narrow when it knows that.

Literals are read with std::from_chars rather than std::stoi/std::stof.
It does not depend on the locale, where std::stof reads "1.5" as 1
wherever ',' is the decimal separator; it does not throw; and it reports
where it stopped. Requiring it to stop at the end of the literal rejects
a malformed literal outright instead of quietly accepting its leading
prefix, which is the same failure the lexer fixes guard against from the
other side.

A missing ')' now throws where it is detected instead of returning null.
parse() caught that at the end anyway, but in between the null travelled
through the tree builders, and every visitor dereferences its children
unguarded.

Parser::cur_token_is is gone. It was declared and defined but never
called from anywhere, while its sibling peek_token_is has seven call
sites.

+++ share/dexpr/include/edp/ast.hpp
@@ -59,8 +59,12 @@ struct StringLiteral {
   std::string value;
 };
 
+// Literals are stored at full precision regardless of how the consumer will
+// eventually use them. Narrowing here would lose information before the
+// expression ever meets a field: 273.15 is 273.14999389648438 as a float, and
+// a legal threshold like 1e40 would not survive at all.
 struct FloatLiteral {
-  float value;
+  double value;
 };
 
 struct IntegerLiteral {
